+++ urls.txt
@@ -1,0 +1,1 @@
+https://raptormaps-imagery.s3.amazonaws.com/upload-data/aefdbb17-b60d-4eff-9f8d-490265b9c07e/DJI_0054.jpg?response-content-type=binary%2Foctet-stream&X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=ASIAXROSU3DWOANTFY7V%2F20260306%2Fus-west-2%2Fs3%2Faws4_request&X-Amz-Date=20260306T190435Z&X-Amz-Expires=10800&X-Amz-SignedHeaders=host&X-Amz-Security-Token=IQoJb3JpZ2luX2VjECIaCXVzLXdlc3QtMiJGMEQCIHRkVT2PPWL5EU%2B5JsFIBX%2BI%2B0sp9kGRI3iEZnZCyvWTAiAVgViHt9bDqgdXfKUowgP8ZLvqXJxrPyLCS%2FeNGJE%2BLirUAwjr%2F%2F%2F%2F%2F%2F%2F%2F%2F%2F8BEAQaDDUxODUyMjE5ODI1MiIMGPdQP4e4aaWKCuJzKqgDb8pHMhWCowBSwxICGazc1IZvs%2BNwVTV2fFS8WXVIg%2BFahsmw1oDap2jRvJanQAV4iOwVfi7nD7Hw9LXZUfJLZeRX2n8BJ9VMxc0biPo%2F9e0yGXmxt27B%2BsL7NljrP6WeUBxStzp6zZ3%2B4ziJzZiIEpUGeTX%2FZXXoLZXgFYgZCRIjyOgjxi8ergZjvlWlWlA1JgimsMB%2BMlCfoeD1CTYMvAU7VpUp4KcOmGuaiIpnMvfX69Nxo9ZiZ%2FOIJle%2FFXWcO3THb1t53CXGTvsnBwDtTyfZtPkUAF7HZ8nLkgZ9sGIgV5Jc0qUL7uQeXSA1vqsIdpwvwYad40D%2BOlH1Pg21ndV7TjXzjHE%2B%2F333Gt%2F87aTqB8zv6VbVM%2BiI8eHk1oMuG5daT6V%2BHYae7EBlSuirNsvdfy9FOpefTKpNcbu%2FRoGWxKYtWtq3gPX0s1fsyWiDng1iRHytQ9kL4GYa2sQW2QyDODEe14b2KJpSEieHknbz6Zbfcn5RXrsSnrKGC44K2czCPCm5uIbPjhgh%2Bt48jxKDxAo4JAoqt06gjMLsq5b5y6AieSqGOjDBq6zNBjqfASVTy334JW%2Bacs2VhQ1DMR0jmvnRwLkgxpeJ0N8q5hOt2zKxyIz%2Bfq7WzZfBm8P78R2ssr97tLBiTxsr9pGEucRrWk1VBpXVZo0vg2dc3U0%2FPmTUKMNd6oKxyBUpf625X%2Bwf4KkTQJk8FL38qivQkD0IUUzrBZ%2Baemw3OruBdqC9Y0CmJ2Ez40plufic49eGlMGjG8A8nxV5cbDubpDTJA%3D%3D&X-Amz-Signature=fb6a654ea33478d60430588d36f95a5d01ed9570058761664883814b5580d8e9
